pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-pipeline-framework"
version = "1.0.0"
description = "GCP Migration Framework - Monorepo root for IDE context help"
requires-python = ">=3.9"
# Note: Internal monorepo packages are not listed as PEP 621 dependencies because they are not
# published to PyPI. They are installed locally via scripts/setup_deployment_venv.sh or by
# installing each subpackage from gcp-pipeline-libraries/ and deployments/ in editable mode.
# Keeping this list empty prevents pip from trying to fetch non-existent distributions.
dependencies = []


[tool.setuptools.packages.find]
where = [
    "gcp-pipeline-libraries/gcp-pipeline-core/src",
    "gcp-pipeline-libraries/gcp-pipeline-beam/src",
    "gcp-pipeline-libraries/gcp-pipeline-orchestration/src",
    "gcp-pipeline-libraries/gcp-pipeline-transform/src",
    "gcp-pipeline-libraries/gcp-pipeline-tester/src",
    "deployments/em-ingestion/src",
    "deployments/loa-ingestion/src"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "apache-beam[gcp]>=2.52.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-pubsub>=2.0.0",
    "apache-airflow>=2.8.0",
    "dbt-bigquery>=1.7.0",
]
